Count the values of a range exactly, with BigInteger
`Discrete.size` and `Discrete.ordinal` returned `long` and saturated at
`Long.MAX_VALUE`, so a domain with more values than a `long` can number
gave up its count, and `Bound.enumerate` fell back to walking the range
to see whether it was too long. They now return `BigInteger`, the count
is always exact, and no range is walked to find out how big it is.

A nine-character word is such a domain: 256^9 = 2^72 values. Ranges
drawn from it, counted from the least word, hold 2^3 and 2^9 words --
made into lists -- and 2^50 and 2^65, which raise `Size`. All four cost
the same to decide.

The walk that stopped at the limit is gone, along with the saturating
arithmetic it existed to cover.

+++ src/main/java/net/hydromatic/morel/eval/Discrete.java
@@ -18,6 +18,7 @@
  */
 package net.hydromatic.morel.eval;
 
+import java.math.BigInteger;
 import java.util.Comparator;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
@@ -57,22 +58,18 @@ public interface Discrete<T> {
   /** Returns the maximum value of this type. */
   T maxValue();
 
-  /**
-   * Returns how many values this type has, saturating at {@link
-   * Long#MAX_VALUE}.
-   */
-  long size();
+  /** Returns how many values this type has. */
+  BigInteger size();
 
   /**
-   * Returns the position of {@code v}, counting from 0 at {@link #minValue},
-   * saturating at {@link Long#MAX_VALUE}.
+   * Returns the position of {@code v}, counting from 0 at {@link #minValue}.
    *
-   * <p>Positions serve only to count the values between two others, so a domain
-   * that has more values than a {@code long} can number may saturate; a caller
-   * that sees {@link Long#MAX_VALUE} learns that the position is too large to
-   * count to, which is all it needs to know.
+   * <p>Positions serve to count the values between two others without visiting
+   * them. A domain can hold far more values than a machine can -- a
+   * nine-character word has 2^72 of them -- so a position is exact rather than
+   * bounded by the width of a machine integer.
    */
-  long ordinal(T v);
+  BigInteger ordinal(T v);
 }
 
 // End Discrete.java
